Rename finance modules for clarity
