+++ public/ncm.html
@@ -62,6 +62,6 @@
     </thead>
     <tbody></tbody>
   </table>
-  <script type="module" src="/src/pages/ncm/main.js"></script>
+  <script type="module" src="/src/pages/ncm/ncmConsole.js"></script>
 </body>
 </html>
