+++ src/config/financeConfig.js
@@ -1,3 +1,4 @@
+// Default finance configuration values
 export const FINANCE = {
   percent_pago_sobre_ml: 0.20,     // pagamos 20% do preço médio ML
   desconto_venda_vs_ml:  0.20,     // vendemos 20% abaixo do preço médio ML
